Refactoring for code generality and comprehensibility.
This commit marks a significant move away from the specifics of
any particular model (though the code is still largely adapted
to a simpler record type, offering a more suitable base for any
future development.

The "team.py" file's only remaining purpose is to hold utility
code, which should be refactored into a separate file or thrown
away as necessary.

Necessary commits for the near future should include the addition
of tests for the basic low-level functions and methods that have so
far simply appeared out of a cloudless sky, and are therefore
highly suspect until properly unit tested.