Refactor existing code to ratonalise the API.
The structure of the code was very much "thrown together",
as a proof-of-concept rather than a properly-architected
project.
The immediate challenge was a lack of separation between
the data-handling framework and the mechanics of the data
objects.
Many functions from other modules
have now migrated to become methods
on the appropriate objects,
and other functions have been moved
to more archtecturally sensible places.